requirements.txt
certifi==2018.4.16
chardet==3.0.4
cookies==2.2.1
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.0.5
django-allauth==0.36.0
django-cors-headers==2.2.0
django-extra-fields==1.0.0
django-filter==1.1.0
django-heroku==0.3.1
django-taggit==0.22.2
django-taggit-serializer==0.1.5
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
drf-friendly-errors==0.13
drf-yasg==1.7.4
future==0.16.0
get==1.0.3
gunicorn==19.8.1
idna==2.6
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
oauthlib==2.1.0
openapi-codec==1.3.2
Pillow==5.1.0
post==1.0.2
psycopg2==2.7.4
psycopg2-binary==2.7.4
public==1.0.3
PyJWT==1.6.3
python3-openid==3.1.0
pytz==2018.4
query-string==1.0.2
request==1.0.2
requests==2.18.4
requests-oauthlib==0.8.0
responses==0.9.0
ruamel.yaml==0.15.37
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
whitenoise==3.3.1
